Updating validations
