--- app/decorators/changeset-route.js
@@ -1,1 +1,0 @@
-export { default } from 'gavant-ember-validations/decorators/changeset-route';
