--- app/decorators/create-changeset.js
@@ -1,1 +1,0 @@
-export { default } from 'gavant-ember-validations/decorators/create-changeset';
