+++ package.json
@@ -23,7 +23,7 @@
     "ember-changeset": "1.3.0",
     "ember-changeset-conditional-validations": "0.2.1",
     "ember-changeset-validations": "1.2.8",
-    "ember-cli-babel": "^6.6.0",
+    "ember-cli-babel": "^7.7.0",
     "ember-cli-htmlbars": "^2.0.1"
   },
   "devDependencies": {
