+++ tests/dummy/app/routes/application.js
@@ -1,19 +1,19 @@
 import Route from '@ember/routing/route';
-import ChangesetRoute from 'gavant-ember-validations/mixins/changeset-route';
+import ChangesetRoute from 'gavant-ember-validations/decorators/changeset-route';
 import { validatePresence } from 'ember-changeset-validations/validators';
 
-const validations = {
+const Validations = {
     name: [validatePresence({presence: true, ignoreBlank: true})],
     radio: [validatePresence({presence: true})]
 };
 
-export default Route.extend(ChangesetRoute, {
-    validations,
-
+@ChangesetRoute
+export default class Application extends Route {
+    validations = Validations;
     model() {
         return {
             name: null,
             radio: null
         };
     }
-});
+}
